Mount /userdata onto the sdcard partition for flip
Seems the normal /userdata likes to corrupt. This should help with that.
MainUI sometimes fails to create needed folders so we make them
automatically in case we're starting up for the first time
on a device with an already corrupt /userdata

NOTE: The alternative to this is to run

e2fsck -f /dev/mtdblock4

on boot to fix the corruption